paternu zalik/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/requirements.in
#
aiomysql==0.2.0
    # via -r requirements/requirements.in
alembic==1.14.0
    # via -r requirements/requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   starlette
bcrypt==4.2.0
    # via passlib
black==24.10.0
    # via -r requirements/requirements.in
build==1.2.2.post1
    # via pip-tools
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.1.7
    # via
    #   black
    #   pip-tools
    #   uvicorn
coverage[toml]==7.6.4
    # via pytest-cov
cryptography==43.0.3
    # via python-jose
ecdsa==0.19.0
    # via python-jose
fastapi==0.115.2
    # via -r requirements/requirements.in
flake8==7.1.1
    # via -r requirements/requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via -r requirements/requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r requirements/requirements.in
mako==1.3.6
    # via alembic
markupsafe==3.0.2
    # via mako
mccabe==0.7.0
    # via flake8
mypy==1.13.0
    # via -r requirements/requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==24.1
    # via
    #   black
    #   build
    #   pytest
passlib[bcrypt]==1.7.4
    # via -r requirements/requirements.in
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements/requirements.in
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   -r requirements/requirements.in
    #   fastapi
pydantic-core==2.23.4
    # via pydantic
pyflakes==3.2.0
    # via flake8
pymysql==1.1.1
    # via aiomysql
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.3
    # via
    #   -r requirements/requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.24.0
    # via -r requirements/requirements.in
pytest-cov==6.0.0
    # via -r requirements/requirements.in
python-dotenv==1.0.1
    # via -r requirements/requirements.in
python-jose[cryptography]==3.3.0
    # via -r requirements/requirements.in
python-multipart==0.0.17
    # via -r requirements/requirements.in
python-slugify==8.0.4
    # via -r requirements/requirements.in
pyyaml==6.0.2
    # via -r requirements/requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy[asyncio]==2.0.36
    # via
    #   -r requirements/requirements.in
    #   alembic
starlette==0.39.2
    # via
    #   -r requirements/requirements.in
    #   fastapi
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.31.1
    # via -r requirements/requirements.in
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
